Implement login/logout automatic logout. Use active session from all cli commmands.
