+++ cli/ApplicationLayer/Commands/Wallet/Login.cs
@@ -11,11 +11,12 @@ using Microsoft.Extensions.DependencyInjection;
 using System;
 using System.Threading.Tasks;
 using BAMWallet.Extensions;
+using BAMWallet.HD;
 
 namespace CLi.ApplicationLayer.Commands.Wallet
 {
     [CommandDescriptor( "login" , "Unlocks wallet and enables wallet commands.")]
-    class Login : LoginBase
+    class Login : Command
     {
         private readonly IConsole _console;
 
@@ -28,7 +29,10 @@ namespace CLi.ApplicationLayer.Commands.Wallet
         public override Task Execute()
         {
             //check if wallet exists, if it does, save session, login and inform command service
-
+            using var identifier = Prompt.GetPasswordAsSecureString("Identifier:", ConsoleColor.Yellow);
+            using var passphrase = Prompt.GetPasswordAsSecureString("Passphrase:", ConsoleColor.Yellow);
+            Session session = new Session(identifier, passphrase); //will throw if wallet doesn't exist
+            Command.ActiveSession = session;
             Login();
             return Task.CompletedTask;
         }
