+++ cli/ApplicationLayer/Commands/Wallet/WalletReceivePaymentCommand.cs
@@ -36,9 +36,7 @@ namespace CLi.ApplicationLayer.Commands.Wallet
 
         public override async Task Execute()
         {
-            using var identifier = Prompt.GetPasswordAsSecureString("Identifier:", ConsoleColor.Yellow);
-            using var passphrase = Prompt.GetPasswordAsSecureString("Passphrase:", ConsoleColor.Yellow);
-
+            Login();
             var paymentId = Prompt.GetString("PAYMENTID:", null, ConsoleColor.Green);
             if (!string.IsNullOrEmpty(paymentId))
             {
