+++ cli/ApplicationLayer/Commands/Wallet/WalletRecoverTransactionsCommand.cs
@@ -24,8 +24,7 @@ namespace CLi.ApplicationLayer.Commands.Wallet
 
         public override async Task Execute()
         {
-            using var identifier = Prompt.GetPasswordAsSecureString("Identifier:", ConsoleColor.Yellow);
-            using var passphrase = Prompt.GetPasswordAsSecureString("Passphrase:", ConsoleColor.Yellow);
+            Login();
             await Spinner.StartAsync("Recovering transactions ...", async spinner =>
             {
                 _spinner = spinner;
